% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.semiopbart.R
\name{adaptive_scheduler}
\alias{adaptive_scheduler}
\title{Adaptive communication scheduler}
\usage{
adaptive_scheduler(
  theta_changes,
  current_every,
  min_every = 5,
  max_every = 200,
  target_change = 0.01,
  momentum = 0.3
)

adaptive_scheduler(
  theta_changes,
  current_every,
  min_every = 5,
  max_every = 200,
  target_change = 0.01,
  momentum = 0.3
)
}
\arguments{
\item{theta_changes}{numeric vector of theta changes from each site}

\item{current_every}{current communication frequency}

\item{min_every}{minimum frequency (most communication)}

\item{max_every}{maximum frequency (least communication)}

\item{target_change}{target relative change threshold}

\item{momentum}{smoothing parameter (0-1)}

\item{theta_history}{list of theta changes from each site}
}
\description{
Adaptive communication scheduler

Adaptive communication scheduler - Fixed
}
