% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.semiopbart.R
\name{.filter_trainable_connections}
\alias{.filter_trainable_connections}
\title{Filter connections to only those with a trainable object}
\usage{
.filter_trainable_connections(conns, train.name = "D_train", verbose = TRUE)
}
\arguments{
\item{conns}{list of DataSHIELD connections}

\item{train.name}{name of the training object to check for}

\item{verbose}{print exclusion messages}
}
\value{
list of connections that have the train object
}
\description{
This is the single shared choke point for all trainers:
- Checks each connection for the presence of `train.name`
- Returns only connections where the object exists
- Logs which sites are excluded (with reason)
- No data is ever returned to the client
}
