% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.semiopbart.R
\name{ds.semiOPBARTWarmupBlock}
\alias{ds.semiOPBARTWarmupBlock}
\title{Warmup block for Architecture D - FIXED}
\usage{
ds.semiOPBARTWarmupBlock(
  conns,
  site_names,
  data.name,
  num_tree,
  k,
  num_iterations,
  rotate_every,
  n_swap,
  theta,
  us,
  p,
  J,
  nfilter_threshold,
  threshold_method,
  state_name,
  n_sites,
  adapt_rate = 0,
  seed = 35
)
}
\description{
Warmup block for Architecture D - FIXED
}
