% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.semiopbart.R
\name{ds.semiOPBARTWarmupBlockE}
\alias{ds.semiOPBARTWarmupBlockE}
\title{Warmup block for Architecture E - FIXED}
\usage{
ds.semiOPBARTWarmupBlockE(
  conns,
  site_names,
  data.name,
  num_tree,
  k,
  num_iterations,
  sync_every,
  theta,
  us,
  p,
  J,
  nfilter_threshold,
  threshold_method,
  state_name,
  adapt_rate = 0,
  seed = 35
)
}
\description{
Warmup block for Architecture E - FIXED
}
