% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.SemiOPBARTExplain.R
\name{semiOPBART_explainThresholds}
\alias{semiOPBART_explainThresholds}
\title{Same idea as semiOPBART_explainCoefficients(), for the ordinal
thresholds (us) instead of the linear coefficients.}
\usage{
semiOPBART_explainThresholds(fit, architecture = c("D", "E", "F"))
}
\description{
Same idea as semiOPBART_explainCoefficients(), for the ordinal
thresholds (us) instead of the linear coefficients.
}
