% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.SemiOPBARTExplain.R
\name{semiOPBART_explainVarImportanceDE}
\alias{semiOPBART_explainVarImportanceDE}
\title{Architectures D/E's variable importance -- wraps the pooled numeric
vector ds.semiOPBARTVarImportance() (ds_semiopbart.R) already returns.}
\usage{
semiOPBART_explainVarImportanceDE(pooled_counts, architecture = c("D", "E"))
}
\arguments{
\item{pooled_counts}{output of ds.semiOPBARTVarImportance()}
}
\description{
Architectures D/E's variable importance -- wraps the pooled numeric
vector ds.semiOPBARTVarImportance() (ds_semiopbart.R) already returns.
}
