% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.SemiOPBARTExplain.R
\name{semiOPBART_plotVarImportance}
\alias{semiOPBART_plotVarImportance}
\title{Bar chart for a semiOPBART_explainVarImportance{F,DE}() data.frame.}
\usage{
semiOPBART_plotVarImportance(
  df,
  title = "Variable importance (mean split count)"
)
}
\description{
Bar chart for a semiOPBART_explainVarImportance{F,DE}() data.frame.
}
